refactor: 문자열 덧셈 계산기, 커스텀 Exception 제거
