--- src/main/java/calculator/exception/CalculatorException.java
@@ -1,9 +1,0 @@
-package calculator.exception;
-
-public class CalculatorException extends RuntimeException {
-
-    public CalculatorException(String message) {
-        super(message);
-    }
-
-}
